Completely reworked chat windows and rooms
Introduced the concept of conversations, which forms the base of
all chat. The window and rooms then end up being merely a UI
component. A lot of things are still broken:

* Invitations and conferencing
* A lot of buttons
* Auto-reconnect of conversations
* Button shortcuts

The filters API has changed completely too. I've changed the code
enough to get it to build and run for IRC. Yet to test on other
protocols.

Yes, this should have been multiple commits. Sorry.