Eliminate PATH_MAX usage
PATH_MAX is not defined on GNU Hurd and they have a good reason for
it:

http://www.gnu.org/software/hurd/community/gsoc/project_ideas/maxpath.html

To summarize, the limits are all different and very vague, with most
implementations choosind to set a limit on path components rather than
the entire path. So avoid it and use your own logic.